✨ Feat: 파일업로드 로직 분리 / 카테고리 분리
